\docType{package}
\name{karskitools}
\alias{karskitools}
\alias{karskitools-package}
\title{karskitools: Collection of functions to facilitate reproducible analysis with R, knitr and pandoc}
\description{
  Selected functions to facilitate reproducible analysis
  with R, knitr and pandoc
}
\details{
  karskitools: Collection of functions to facilitate
  reproducible analysis with R, knitr and pandoc

  \tabular{ll}{ Package: \tab karskitools\cr Type: \tab
  Package\cr Version: \tab See sessionInfo() or DESCRIPTION
  file\cr Date: \tab 2013\cr License: \tab MIT\cr LazyLoad:
  \tab yes\cr }
}
\examples{
# library(karskitools)
}
\author{
  Markus Kainu <markuskainu(at)gmail.com>
  \url{http://markuskainu.fi/karskitools}
  \url{https://github.com/muuankarski/karskitools}
}

