\name{knitpandoc_docx}
\alias{knitpandoc_docx}
\title{Convert R Markdown file into .docx format}
\usage{
  knitpandoc_docx(filename, ending = "empty")
}
\arguments{
  \item{filename}{R Markdown file without filename
  extension .Rmd}
}
\value{
  A .docx-format document
}
\description{
  Convert R Markdown file into .docx format
}
\examples{
# see http://markuskainu.fi/karskitools
}
\author{
  Markus Kainu <markuskainu(at)gmail.com>
}

